Add datapath to FlowModMsgBuilder
With more switches in the fabric we cannot infer the switch
from the rule type. This commit adds a new field to
the FlowModMsg builder so we can pass the name of the datapath
as an argument. The default of the method add_flow_mod is None,
so we keep the backward compatibility with the Multi-Table and
Multi-Switch examples.

It also modifies the parsing of the configuration
of a participant port. Now it also contains
the edge switch connected to the participant port.